Rename HAL file to reflect more exact microcontroller model.
